refactor(backend): move  function that use argon2
