+++ server/password/argon2/argon2_test.go
@@ -1,4 +1,4 @@
-package password
+package argon2
 
 import (
 	"github.com/stretchr/testify/assert"
